+++ src/configuration/from/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+    commandLine: require('./commandLine'),
+    //defaults: require('./defaults'),
+    environment: require('./environment'),
+    file: require('./file'),
+    settingsJson: require('./settingsJson')
+}
