+++ test/data/sql/infrastructure/config.js
@@ -6,4 +6,4 @@
 var configuration = require('../../../../src/configuration'),
     path = require('path');
 
-module.exports = configuration.fromEnvironment(configuration.fromFile({}, path.resolve(__dirname, '../../../config.js'))).data;
+module.exports = configuration.from.environment(configuration.from.file({}, path.resolve(__dirname, '../../../config.js'))).data;
